+++ Sources/PeripheryKit/Analyzer/Visitors/ComplexPropertyAccessorReferenceBuilder.swift
@@ -3,8 +3,7 @@ import Foundation
 // Builds references to getters and setters from complex properties. A complex property is one that
 // implements a get, set, willSet or didSet. Accessors are distinct declarations and hold references,
 // rather than the property declaration itself. References are not built for simple property
-// accessors as they do not hold references and it aids identification of unread properties in
-// aggressive mode.
+// accessors as they do not hold references.
 final class ComplexPropertyAccessorReferenceBuilder: SourceGraphVisitor {
     static func make(graph: SourceGraph) -> Self {
         return self.init(graph: graph)